Move util func to connection-provider-resolver
