+++ packages/server/api/src/app/app-connection/connection-providers-resolver.ts
@@ -92,3 +92,25 @@ export async function getAuthProviderMetadata(
   return blocks.find((block) => block.auth?.authProviderKey === authProviderKey)
     ?.auth;
 }
+
+export async function findBlockByAuthProviderKey(
+  authProviderKey: string,
+  projectId: string,
+): Promise<{ name: string; version: string } | undefined> {
+  const release = await flagService.getCurrentRelease();
+  const edition = system.getEdition();
+
+  const blocks = await blockMetadataService.list({
+    includeHidden: false,
+    projectId,
+    release,
+    edition,
+  });
+
+  const block = blocks.find(
+    (block) => block.auth?.authProviderKey === authProviderKey,
+  );
+
+  if (!block) return undefined;
+  return { name: block.name, version: block.version };
+}
